move neo4j tests to main test folder. Comment out warnings
