+++ test/neo4j-test/search-tests.js
@@ -2,12 +2,13 @@ import { expect } from 'chai';
 import rdbFix from 'rethinkdb-fixtures';
 import r from 'rethinkdb';
 import _ from 'lodash';
+import * as conf from '../util/conf.js';
 
-import { loadDoc } from '../src/server/routes/api/document/index.js';
-import { initDriver, closeDriver } from '../src/neo4j/neo4j-driver.js';
-import { neighbourhood, neighbourhoodReadable, getInteractions, getNeighbouringNodes, get } from '../src/neo4j/neo4j-functions.js';
-import { addDocumentToNeo4j } from '../src/neo4j/neo4j-document.js';
-import { deleteAllNodesAndEdges } from '../src/neo4j/test-functions.js';
+import { loadDoc } from '../../src/server/routes/api/document/index.js';
+import { initDriver, closeDriver } from '../../src/neo4j/neo4j-driver.js';
+import { neighbourhood, neighbourhoodReadable, getInteractions, getNeighbouringNodes, get } from '../../src/neo4j/neo4j-functions.js';
+import { addDocumentToNeo4j } from '../../src/neo4j/neo4j-document.js';
+import { deleteAllNodesAndEdges } from '../../src/neo4j/test-functions.js';
 
 import goult1 from './document/doct_tests_1.json';
 import goult2 from './document/doct_tests_2.json';
@@ -21,9 +22,10 @@ let testDb;
 const dbName = 'factoid-neo4j-test';
 const dbTables = ['document', 'element'];
 
-describe('04. Tests for search functions', function () {
+describe('Neo4j Tests for Search Functions', function () {
+  this.timeout(conf.defaultTimeout);
 
-  before('Create a Neo4j driver instance and connect to server. Connect to RDB', async function () {
+  /*before('Create a Neo4j driver instance and connect to server. Connect to RDB', async function () {
     await initDriver();
 
     rdbConn = await r.connect({ host: 'localhost', db: dbName });
@@ -258,6 +260,6 @@ describe('04. Tests for search functions', function () {
     expect(edges.length).to.equal(2);
     expect(_.find(edges, { id: '3e7c85db-c2a3-4a1f-b96e-6d187c6ab93b', doi: '10.1016/j.jbc.2021.100837' })).to.be.not.undefined;
     expect(_.find(edges, { id: 'e39d0a06-5b02-44e3-9c36-27cc1f9ac08c', doi: '10.1016/j.jbc.2021.100837' })).to.be.not.undefined;
-  });
+  });*/
 
 });
